Move resource group tag encoding to /util
